pyproject.toml
[tool.poetry]
name = "fintrack-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.2"
fastapi = "^0.115.0"
jwt = "^1.3.1"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
